requirements.txt
streamlit
numpy
faiss-cpu
openai
tqdm